src/lambda_requirements.txt
requests
psycopg2-binary
neo4j